BUGFIX in UPTREND indicator builder

+++ indicators_builder.py
@@ -7,8 +7,8 @@ from datetime import datetime
 
 
 def do_uptrend(source: pd.DataFrame, name):
-	uptrend = pd.Series((source[name][1:].reset_index(drop=True).astype('int') >
-						 source[name][:-1].reset_index(drop=True).astype('int')), name=name + "_UP")
+	uptrend = pd.Series((source[name][1:].reset_index(drop=True) >
+						 source[name][:-1].reset_index(drop=True)).astype('int'), name=name + "_UP")
 
 	source = source.join(uptrend)
